Changes to add a supersource class and use it from a client type
Step 2: created a java.awt.Color supersource, with a subset of the
original structure. Located directly in src/main/super, step 5 will
ensure that this is relocated to match the super-source tag (as set in
step 4).

Step 3: replace the contents of ColorWeb with just an import to
java.awt.Color and invocation of its constructor.

Step 4: add super-source tag to module.gwt.xml. This is necessary
because the file already contains a <source> tag, otherwise both would
be auto-generated.

Step 5: specify relocateSuperSource=true in the gwt-maven-plugin. Note
that this is optional - if Color was in
src/main/super/com/company/super/java/awt/, this step would not be
needed.

To verify this commit, run mvn verify from the command line.